refactor(python): extract model websocket usage lifecycle
