+++ crates/runner/mitm-addon/tests/model_provider_websocket_helpers.py
@@ -9,7 +9,7 @@ from wsproto.frame_protocol import Opcode
 
 import deferred_callbacks
 import mitm_addon
-import response_streaming
+import model_websocket_usage
 
 _WebSocketTrimCallback = Callable[[http.HTTPFlow], None]
 ScheduledWebSocketTrim = tuple[_WebSocketTrimCallback, http.HTTPFlow]
@@ -91,7 +91,7 @@ def set_websocket_message(
 
 
 def _assert_model_websocket_usage_started(flow: http.HTTPFlow) -> None:
-    assert response_streaming.is_model_websocket_usage_enabled(flow)
+    assert model_websocket_usage.is_enabled(flow)
 
 
 def feed_websocket_server_message(flow: http.HTTPFlow, content: bytes) -> None:
